requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1
python-dotenv==1.0.1
opencv-python==4.10.0.84
torch==2.3.1
torchvision==0.18.1
ultralytics==8.3.0
numpy==1.26.4
scikit-learn
matplotlib
Pillow
tensorboard
pytorchvideo